refactor individual component * create header component * create subscription component * create detail pane ** all values as observables ** last observation values are taken from data and not processed in the app no more * create observation view * create edit headers * create individual-edit-view * reuse detail-header in stations * remove base individual detail class * restructured component folders
