+++ src/app/individual/individual-description/individual-description.component.html
@@ -1,0 +1,47 @@
+<div class="detail-label">{{ (species | async)?.de | translate }}</div>
+<div class="title-individual">{{ (individual | async).name }}</div>
+<div class="detail-label" *ngIf="(individual | async).last_observation_date">
+    <span translate>Letzte Beobachtung</span>
+    <span>, {{ lastObservationDate | async }}</span>
+</div>
+<div class="phenophase-title" *ngIf="lastPhenophase | async">
+    <span class="phenophase-circle" [ngStyle]="{ 'background-color': lastPhenophaseColor | async }"></span>
+    <span>{{ (lastPhenophase | async)?.name_de | translate }}</span>
+</div>
+<div>
+    <span>{{ 'Von' | translate }} </span
+    ><a mat-button [routerLink]="['/profile', (individual | async).user]" class="creator">{{
+    individualCreatorNickname | async
+    }}</a>
+</div>
+<div class="detail-label">{{ 'Umgebung' | translate }}</div>
+<div class="detail-info-value">{{ (description | async)?.de | translate }}</div>
+<div class="detail-label">{{ 'Höhe' | translate }}</div>
+<div>
+	<span class="detail-info-value">{{ (individual | async)?.altitude }}</span>
+	<span translate> {{ 'M.ü.M.' }}</span>
+</div>
+<div class="detail-label">{{ 'Exposition' | translate }}</div>
+<div class="detail-info-value">{{ (exposition | async)?.de | translate }}</div>
+<div class="detail-label">{{ 'Hangneigung' | translate }}</div>
+<div class="detail-info-value">{{ (individual | async)?.gradient }}%</div>
+<div class="detail-label">{{ 'Schatten' | translate }}</div>
+<div class="detail-info-value">{{ (shade | async)?.de | translate }}</div>
+<div class="detail-label">{{ 'Bewässerung' | translate }}</div>
+<div class="detail-info-value">{{ (irrigation | async)?.de | translate }}</div>
+<div class="detail-label">{{ 'Distanz zu Gebäude / Strasse' | translate }}</div>
+<div class="detail-info-value">{{ (distance | async)?.de | translate }}</div>
+<div class="detail-label">{{ 'Habitat' | translate }}</div>
+<div class="detail-info-value">{{ (habitat | async)?.de | translate }}</div>
+<div class="detail-label">{{ 'Waldtyp' | translate }}</div>
+<div class="detail-info-value">{{ (forest | async)?.de | translate }}</div>
+
+<button
+	mat-flat-button
+	color="primary"
+	type="submit"
+	id="edit-button"
+	[routerLink]="['/individuals', individualId, 'edit']"
+	*ngIf="isEditable | async">
+  {{ 'Objekt bearbeiten' | translate }}
+</button>
