+++ src/app/individual/individual-description/individual-description.component.scss
@@ -1,0 +1,28 @@
+.title-individual {
+  font-size: 1.375rem;
+  letter-spacing: 0.055rem;
+  line-height: 1.875rem;
+}
+
+.detail-info-value {
+  font-size: 0.875rem;
+  font-weight: 500;
+  letter-spacing: 0.07rem;
+  color: #5d5d5d;
+}
+
+.creator {
+  font-weight: 700;
+}
+
+#edit-button {
+  margin-top: 4.125rem;
+  padding-left: 40px;
+  padding-right: 40px;
+}
+
+.phenophase-title {
+  font-size: 1.375rem;
+  padding-top: 0.125rem;
+  padding-bottom: 1.25rem;
+}
