--- src/app/individual/individual-detail.component.html
@@ -1,143 +1,0 @@
-<div gdColumns="repeat(8, 1fr)">
-  <div gdColumn="2 / span 6" gdColumn.lt-md="1 / span 8" gdColumns="repeat(6, 1fr)" class="detail-view">
-    <div gdColumn="1 / span 6" class="detail-header">
-      <div gdColumns="repeat(6, 1fr)">
-        <google-map
-          [class]="(imageUrl | async) ? 'individual-map-with-image' : 'individual-map'"
-          [gdColumn]="(imageUrl | async) ? '1 / span 4' : '1 / span 6'"
-          width="100%"
-          height="302px"
-          [center]="center"
-          [zoom]="zoom"
-          [options]="options"
-        >
-          <map-marker #marker [position]="geopos" [options]="markerOptions | async"></map-marker>
-        </google-map>
-        <div gdColumn="5 / span 2" class="image-container" *ngIf="imageUrl | async; let url">
-          <img [src]="url" class="individual-image" width="100%" />
-        </div>
-      </div>
-      <button
-        *ngIf="isLoggedIn && !isOwner() && isFollowing | async"
-        mat-fab
-        class="detail-header-fab a100"
-        color="primary"
-        (click)="unfollow()"
-      >
-        <mat-icon svgIcon="subscribe-active"></mat-icon>
-      </button>
-      <button
-        *ngIf="isLoggedIn && !isOwner() && !(isFollowing | async)"
-        mat-fab
-        class="detail-header-fab"
-        color="primary"
-        (click)="follow()"
-      >
-        <mat-icon svgIcon="subscribe"></mat-icon>
-      </button>
-    </div>
-    <div gdColumn="1 / span 6" gdColumns="repeat(6, 1fr)" class="detail-content">
-      <div gdColumn="1 / span 2" gdColumn.lt-md="1 / span 6" *ngIf="detailSubject | async" class="detail-left">
-        <div class="detail-label">{{ (species | async)?.de | translate }}</div>
-        <div class="title-individual">{{ (detailSubject | async).name }}</div>
-        <div class="detail-label" *ngIf="(detailSubject | async).last_observation_date">
-          <span translate>Letzte Beobachtung</span>
-          <!-- TODO remove method call -->
-          <span>, {{ lastObservationDate }}</span>
-        </div>
-        <div class="phenophase-title" *ngIf="lastPhenophase | async">
-          <span class="phenophase-circle" [ngStyle]="{ 'background-color': lastPhenophaseColor | async }"></span>
-          <span>{{ (lastPhenophase | async)?.name_de | translate }}</span>
-        </div>
-        <div>
-          <span>{{ 'Von' | translate }} </span
-          ><a mat-button [routerLink]="['/profile', (detailSubject | async).user]" class="creator">{{
-            individualCreatorNickname | async
-          }}</a>
-        </div>
-        <div class="detail-label">{{ 'Umgebung' | translate }}</div>
-        <div class="detail-info-value">{{ (description | async)?.de | translate }}</div>
-        <div class="detail-label">{{ 'Höhe' | translate }}</div>
-        <div>
-          <span class="detail-info-value">{{ (detailSubject | async)?.altitude }}</span
-          ><span translate> {{ 'M.ü.M.' }}</span>
-        </div>
-        <div class="detail-label">{{ 'Exposition' | translate }}</div>
-        <div class="detail-info-value">{{ (exposition | async)?.de | translate }}</div>
-        <div class="detail-label">{{ 'Hangneigung' | translate }}</div>
-        <div class="detail-info-value">{{ (detailSubject | async)?.gradient }}%</div>
-        <div class="detail-label">{{ 'Schatten' | translate }}</div>
-        <div class="detail-info-value">{{ (shade | async)?.de | translate }}</div>
-        <div class="detail-label">{{ 'Bewässerung' | translate }}</div>
-        <div class="detail-info-value">{{ (irrigation | async)?.de | translate }}</div>
-        <div class="detail-label">{{ 'Distanz zu Gebäude / Strasse' | translate }}</div>
-        <div class="detail-info-value">{{ (distance | async)?.de | translate }}</div>
-        <div class="detail-label">{{ 'Habitat' | translate }}</div>
-        <div class="detail-info-value">{{ (habitat | async)?.de | translate }}</div>
-        <div class="detail-label">{{ 'Waldtyp' | translate }}</div>
-        <div class="detail-info-value">{{ (forest | async)?.de | translate }}</div>
-
-        <button mat-flat-button color="primary" type="submit" id="edit-button" (click)="editDetail()" *ngIf="isOwner()">
-          {{ 'Objekt bearbeiten' | translate }}
-        </button>
-      </div>
-      <div gdColumn="3 / span 4" gdColumn.lt-md="1 / span 6" class="detail-right">
-        <h3>{{ 'Phänophasen' | translate }}</h3>
-        <div *ngFor="let groupedItem of phenophaseObservationsGroups | async" class="phenophase-group">
-          <ng-container *ngIf="isOwner() || groupedItem.hasObservations">
-            <h4>{{ groupedItem.phenophaseGroup.name_de | translate }}</h4>
-            <div *ngFor="let item of groupedItem.phenophaseObservations">
-              <ng-container *ngIf="isOwner() || item.observation.isSome()">
-                <div class="detail-label">{{ item.phenophase.name_de | translate }}</div>
-                <mat-form-field class="full-width detail-observation-date">
-                  <input
-                    matInput
-                    [matDatepicker]="fakePicker"
-                    [id]="item.id"
-                    placeholder="{{ item.phenophase.description_de | translate }}"
-                    [value]="item.observation.toNullable()?.date"
-                    readonly="true"
-                  />
-                  <mat-icon
-                    *ngIf="isOwner() && item.observation.isNone() && item.limits.isEditable"
-                    matSuffix
-                    class="add-phenophase"
-                    color="primary"
-                    (click)="editPhenophaseDate(item)"
-                    >add_circle</mat-icon
-                  >
-                  <mat-icon
-                    *ngIf="isOwner() && item.observation.isNone() && !item.limits.isEditable"
-                    matSuffix
-                    class="add-phenophase-disabled"
-                    color="accent"
-                    >add_circle</mat-icon
-                  >
-                  <mat-icon
-                    *ngIf="isOwner() && item.observation.isSome()"
-                    matSuffix
-                    color="primary"
-                    class="edit-phenophase"
-                    (click)="editPhenophaseDate(item)"
-                    >edit</mat-icon
-                  >
-                  <mat-icon
-                  *ngIf="isOwner() && item.observation.isSome()"
-                  matSuffix
-                  color="warn"
-                  class="edit-phenophase"
-                  (click)="deletePhenophaseDate(item)"
-                  >delete</mat-icon
-                >
-                </mat-form-field>
-                <div class="detail-sub" *ngIf="item.observation.toNullable()?.comment">
-                  {{ 'Kommentar' }}: {{ staticComments[item.observation.toNullable()?.comment] | translate }}
-                </div>
-              </ng-container>
-            </div>
-          </ng-container>
-        </div>
-      </div>
-    </div>
-  </div>
-</div>
