--- src/app/individual/individual-detail.component.scss
@@ -1,99 +1,0 @@
-@import '../../theming/palettes';
-
-.geopos-width {
-  width: 70%;
-}
-
-.creator {
-  font-weight: 700;
-}
-
-#edit-button {
-  margin-top: 4.125rem;
-  padding-left: 40px;
-  padding-right: 40px;
-}
-
-.phenophase-title {
-  font-size: 1.375rem;
-  padding-top: 0.125rem;
-  padding-bottom: 1.25rem;
-}
-
-.phenophase-group {
-  margin-top: 2.5625rem;
-}
-
-h4 {
-  color: #868072;
-}
-
-.detail-header-fab {
-  margin-left: 90%;
-  bottom: 28px;
-
-  width: 70px;
-  height: 70px;
-}
-
-@media screen and (max-width: 959px) {
-  .detail-header-fab {
-    margin-left: 70%;
-  }
-}
-
-.detail-sub {
-  font-style: italic;
-  margin-top: -1.25rem;
-  padding-bottom: 1.25rem;
-}
-
-.following {
-  color: #1dff6e; //mat-color($phaenonet-p, A700);
-}
-
-.add-phenophase {
-  cursor: pointer;
-}
-
-.add-phenophase-disabled {
-  cursor: not-allowed;
-  color: #f7f5f2;
-}
-
-.edit-phenophase {
-  cursor: pointer;
-}
-
-.image-container {
-  width: 100%;
-  height: 302px;
-  overflow: hidden;
-}
-
-.individual-image {
-  height: 302px;
-  object-fit: cover;
-}
-
-.individual-map-with-image ::ng-deep .map-container {
-  border-top-left-radius: 1rem;
-  border-top-right-radius: 0rem;
-}
-
-.detail-view {
-  background-color: white;
-}
-
-.title-individual {
-  font-size: 1.375rem;
-  letter-spacing: 0.055rem;
-  line-height: 1.875rem;
-}
-
-.detail-info-value {
-  font-size: 0.875rem;
-  font-weight: 500;
-  letter-spacing: 0.07rem;
-  color: #5d5d5d;
-}
