+++ src/app/individual/individual-detail/individual-detail-header/individual-detail-header.component.html
@@ -1,0 +1,15 @@
+<div gdColumns="repeat(6, 1fr)">
+    <google-map
+        [class]="(imageUrl | async) ? 'individual-map-with-image' : 'individual-map'"
+        [gdColumn]="(imageUrl | async) ? '1 / span 4' : '1 / span 6'"
+        width="100%"
+        height="302px"
+        [center]="center | async"
+        [zoom]="zoom"
+        [options]="options">
+        <map-marker #marker [position]="geopos | async" [options]="markerOptions | async"></map-marker>
+    </google-map>
+    <div gdColumn="5 / span 2" class="image-container" *ngIf="imageUrl | async; let url">
+        <img [src]="url" class="individual-image" width="100%" />
+    </div>
+</div>
