+++ src/app/individual/individual-detail/individual-detail-header/individual-detail-header.component.scss
@@ -1,0 +1,15 @@
+.individual-image {
+  height: 302px;
+  object-fit: cover;
+}
+
+.individual-map-with-image ::ng-deep .map-container {
+  border-top-left-radius: 1rem;
+  border-top-right-radius: 0rem;
+}
+
+.image-container {
+  width: 100%;
+  height: 302px;
+  overflow: hidden;
+}
