+++ src/app/individual/individual-detail/individual-detail.component.scss
@@ -1,0 +1,3 @@
+.detail-view {
+  background-color: white;
+}
