+++ src/app/individual/individual-detail/individual-subscription/individual-subscription.component.html
@@ -1,0 +1,20 @@
+<div>
+    <button
+      *ngIf="isFollowing | async"
+      mat-fab
+      class="detail-header-fab a100"
+      color="primary"
+      (click)="unfollow()"
+    >
+      <mat-icon svgIcon="subscribe-active"></mat-icon>
+    </button>
+    <button
+      *ngIf="!(isFollowing | async)"
+      mat-fab
+      class="detail-header-fab"
+      color="primary"
+      (click)="follow()"
+    >
+      <mat-icon svgIcon="subscribe"></mat-icon>
+    </button>
+</div>
