+++ src/app/individual/individual-detail/individual-subscription/individual-subscription.component.scss
@@ -1,0 +1,13 @@
+.detail-header-fab {
+  margin-left: 90%;
+  bottom: 28px;
+
+  width: 70px;
+  height: 70px;
+}
+
+@media screen and (max-width: 959px) {
+  .detail-header-fab {
+    margin-left: 70%;
+  }
+}
