+++ src/app/individual/individual-edit/individual-edit-header/individual-edit-header.component.scss
@@ -1,29 +1,3 @@
-.geopos-width {
-  width: 70%;
-}
-
-@media screen and (max-width: 959px) {
-  .geopos-width {
-    width: 100%;
-  }
-}
-
-.right-align {
-  text-align: right;
-}
-
-.gradient-info-img {
-  height: 61px;
-  padding-left: 2rem;
-}
-
-@media screen and (max-width: 959px) {
-  .gradient-info-img {
-    height: 50px;
-    padding-left: 1rem;
-  }
-}
-
 .image-container {
   width: 100%;
   height: 402px;
@@ -38,14 +12,6 @@
 .individual-map-with-image ::ng-deep .map-container {
   border-top-left-radius: 1rem;
   border-top-right-radius: 0rem;
-}
-
-.detail-view {
-  background-color: white;
-}
-
-.add-image {
-  cursor: pointer;
 }
 
 .detail-header-fab {
