+++ src/app/individual/individual-edit/individual-edit-header/individual-edit-header.component.ts
@@ -1,0 +1,64 @@
+import { Component, Input, OnInit } from '@angular/core';
+import { AngularFireAnalytics } from '@angular/fire/analytics';
+import { Observable, ReplaySubject } from 'rxjs';
+import { first, map, mergeAll } from 'rxjs/operators';
+import { Individual } from '../../individual';
+import { IndividualService } from '../../individual.service';
+import { GeoposService } from './geopos.service';
+
+@Component({
+  selector: 'app-individual-edit-header',
+  templateUrl: './individual-edit-header.component.html',
+  styleUrls: ['./individual-edit-header.component.scss']
+})
+export class EditHeaderComponent implements OnInit {
+  @Input() individual: ReplaySubject<Individual>;
+
+  initialGeopos = this.geoposService.initialGeopos;
+  zoom = 9;
+  options: google.maps.MapOptions = {
+    mapTypeId: google.maps.MapTypeId.HYBRID,
+    mapTypeControl: false,
+    fullscreenControl: false,
+    streetViewControl: false,
+    minZoom: 8
+  };
+  markerOptions: google.maps.MarkerOptions = {
+    draggable: true,
+    icon: { url: '/assets/img/map_pins/map_pin_generic.png', scaledSize: new google.maps.Size(55, 60) }
+  };
+
+  imageUrl: Observable<string>;
+
+  constructor(
+    private geoposService: GeoposService,
+    private individualService: IndividualService,
+    private analytics: AngularFireAnalytics
+  ) {}
+
+  ngOnInit() {
+    this.individual.pipe(first()).subscribe(i => {
+      if (i.geopos) {
+        this.geoposService.update(new google.maps.LatLng(i.geopos));
+      }
+    });
+    this.imageUrl = this.individual.pipe(map(individual =>
+      this.individualService.getImageUrl(individual, true)
+      ), mergeAll());
+  }
+
+  updateGeopos(event: google.maps.MouseEvent): void {
+    this.geoposService.update(event.latLng);
+  }
+
+  locateMe(): void {
+    if (navigator.geolocation) {
+      navigator.geolocation.getCurrentPosition(position => {
+        this.geoposService.update(
+          new google.maps.LatLng({ lat: position.coords.latitude, lng: position.coords.longitude })
+        );
+      });
+    }
+    this.analytics.logEvent('individual.locate-me');
+  }
+}
