+++ src/app/individual/individual-edit/individual-edit.component.html
@@ -1,0 +1,19 @@
+<div gdColumns="repeat(8, 1fr)">
+  <div gdColumn="2 / span 6" gdColumn.lt-md="1 / span 8" gdColumns="repeat(6, 1fr)" class="detail-view">
+    <app-individual-edit-header
+      gdColumn="1 / span 6"
+      class="detail-header"
+      [individual]='detailSubject'></app-individual-edit-header>
+    <div gdColumn="1 / span 6" gdColumns="repeat(6, 1fr)" class="detail-content">
+      <div gdColumn="1 / span 2" gdColumn.lt-md="1 / span 6" class="detail-left">
+        <h3 translate>OBJEKT ERFASSEN</h3>
+      </div>
+      <app-individual-edit-view
+        gdColumn="3 / span 4"
+        gdColumn.lt-md="1 / span 6"
+        class="detail-right"
+        [individual]='detailSubject'
+        [createNewIndividual]='isNewIndividual'></app-individual-edit-view>
+    </div>
+  </div>
+</div>
