+++ src/app/individual/individual-edit/individual-edit.component.scss
@@ -1,0 +1,3 @@
+.detail-view {
+  background-color: white;
+}
