+++ src/app/individual/individual-routing.module.ts
@@ -1,8 +1,8 @@
 import { AuthGuard } from './../auth/auth-guard.service';
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
-import { IndividualEditComponent } from './individual-edit.component';
-import { IndividualDetailComponent } from './individual-detail.component';
+import { IndividualEditComponent } from './individual-edit/individual-edit.component';
+import { IndividualDetailComponent } from './individual-detail/individual-detail.component';
 
 const routes: Routes = [
   {
