+++ src/app/station/station-detail.component.html
@@ -1,28 +1,14 @@
 <div gdColumns="repeat(8, 1fr)">
   <div gdColumn="2 / span 6" gdColumn.lt-md="1 / span 8" gdColumns="repeat(6, 1fr)" class="detail-view">
-    <div gdColumn="1 / span 6" class="detail-header">
-      <google-map class="map" width="100%" height="302px" [center]="center" [zoom]="zoom" [options]="options">
-        <map-marker #marker [position]="geopos" [options]="markerOptions | async"></map-marker>
-      </google-map>
-      <button
-        *ngIf="isLoggedIn && isFollowing | async"
-        mat-fab
-        class="detail-header-fab a100"
-        color="primary"
-        (click)="unfollow()"
-      >
-        <mat-icon svgIcon="subscribe-active"></mat-icon>
-      </button>
-      <button
-        *ngIf="isLoggedIn && !(isFollowing | async)"
-        mat-fab
-        class="detail-header-fab"
-        color="primary"
-        (click)="follow()"
-      >
-        <mat-icon svgIcon="subscribe"></mat-icon>
-      </button>
-    </div>
+    <app-individual-detail-header
+      gdColumn="1 / span 6"
+      class="detail-header"
+      [individual]="detailSubject"></app-individual-detail-header>
+    <app-individual-subscription
+      *ngIf="isLoggedIn"
+      gdColumn="1 / span 6"
+      class="detail-header"
+      [individual]="detailSubject"></app-individual-subscription>
     <div gdColumn="1 / span 6" gdColumns="repeat(6, 1fr)" class="detail-content">
       <div gdColumn="1 / span 2" gdColumn.lt-md="1 / span 6" *ngIf="detailSubject | async" class="detail-left">
         <div class="detail-label detail-label-standort">{{ 'Standort' | translate }}</div>
